Split out test classes to avoid potential racing conditions
